+++ CHANGELOG.md
@@ -5,11 +5,14 @@
 ### New Features
 
 - Added support for creating vectorized UDTFs with `process` method.
+- Added support for `locate` in `snowflake.snowpark.functions`.
 - Added support for dataframe functions:
   - to_timestamp_ltz
   - to_timestamp_ntz
   - to_timestamp_tz
-- Added support for the following local testing functions:
+- Added support for ASOF JOIN type.
+- Added support for the following local testing APIs:
+  - to_double
   - to_timestamp
   - to_timestamp_ltz
   - to_timestamp_ntz
@@ -17,8 +20,7 @@
   - greatest
   - least
   - dateadd
-- Added support for ASOF JOIN type.
-- Added support for the following local testing APIs:
+  - date_part
   - Session.get_current_account
   - Session.get_current_warehouse
   - Session.get_current_role